% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markback.R
\name{marksback}
\alias{marksback}
\title{Process Markdown files containing student grades and feedback.}
\usage{
marksback(markdown, savegrades = T, savefeedback = T,
  output_folder = "processed")
}
\description{
Process Markdown files containing student grades and feedback.
}
